Remove typescript from frontend
