--- model/TimetableConfiguration.ts
@@ -1,5 +1,0 @@
-export default class TimetableConfiguration {
-  id: string[];
-  label: string;
-  whitelist: string[];
-}
